# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "nadeemshaik1171998@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/createNotes": {
            "post": {
                "summary": "create note",
                "description": "create user note",
                "operationId": "8aa521064947715a5fdc7f892ef0117a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "notes created successfully"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/displayNoteById": {
            "get": {
                "summary": "display Note",
                "description": "user display Note",
                "operationId": "e3ed7f34d259acd3b79483130b54bd56",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "All Notes are Fetched Successfully"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/updateNoteById": {
            "post": {
                "summary": "update note",
                "description": "update user note",
                "operationId": "d5fa69313342590d7ae2dd4b243a1d65",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "id",
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "id": {},
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note successfully updated"
                    },
                    "404": {
                        "description": "Notes not found"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/deleteNoteById": {
            "post": {
                "summary": "delete note",
                "description": "delete user note",
                "operationId": "6fa2f7be0614f8bfc72ad4a90d037160",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note deleted Successfully"
                    },
                    "404": {
                        "description": "Notes not found"
                    },
                    "401": {
                        "description": "Invalid Authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/paginationNote": {
            "get": {
                "summary": "Display Paginate Notes",
                "description": " Display Paginate Notes ",
                "operationId": "3b08f9755ac9b6beaabc79a0a282b9da",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "Pagination aplied to all Notes"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/pinNoteById": {
            "post": {
                "summary": "Pin Note",
                "description": " Pin Note ",
                "operationId": "2e1d7173b89fa51f02111ed713e89efa",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note Pinned Sucessfully"
                    },
                    "404": {
                        "description": "Notes not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/getAllPinnedNotes": {
            "get": {
                "summary": "Display All Pinned Notes",
                "description": " Display All Pinned Notes ",
                "operationId": "b0009c3aaede3bdd703897789891b65b",
                "requestBody": {},
                "responses": {
                    "404": {
                        "description": "Invalid token"
                    },
                    "201": {
                        "description": "Fetched Pinned Notes Successfully"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/archiveNoteById": {
            "post": {
                "summary": "Archive Note",
                "description": " Archive Note ",
                "operationId": "d161079104ec20f554ac3a10c5e01394",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note Archived Sucessfully"
                    },
                    "404": {
                        "description": "Notes not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/getAllArchiveNotes": {
            "get": {
                "summary": "Display All Archived Notes",
                "description": " Display All Archived Notes ",
                "operationId": "541c9c4c7b39fc32d61aa29831c8a2ae",
                "requestBody": {},
                "responses": {
                    "404": {
                        "description": "Invalid token"
                    },
                    "201": {
                        "description": "Fetched Archived Notes"
                    },
                    "403": {
                        "description": "Notes not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/colourNoteById": {
            "post": {
                "summary": "Colour Note",
                "description": " Colour Note ",
                "operationId": "8e86cf0f3b2c3fa1727daa2f31814261",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "colour"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "colour": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note coloured Sucessfully"
                    },
                    "404": {
                        "description": "Notes not Found"
                    },
                    "400": {
                        "description": "Colour Not Specified in the List"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}